requirements.txt
PySide6
requests
pynvml
psutil
py-cpuinfo
PySide6-Addons
PyGithub